backend/requirements.txt
fastapi==0.105.0
python-multipart==0.0.6
pillow==10.1.0
tensorflow==2.15.0  # Changed from tensorflow-macos
uvicorn==0.24.0